requirements.txt
langchain==1.2.18
langchain_classic==1.0.7
langchain_community==0.4.1
langchain_core==1.3.3
langchain_openai==1.2.1
python-dotenv==1.2.2
streamlit==1.57.0
chromadb
rank_bm25
langgraph
langgraph
